Update: 10 Februari 2022 - Nilai change to float number - Show nilai for kerja praktek

+++ database/migrations/2021_10_16_023319_create_table_form_penilaian_item.php
@@ -18,8 +18,8 @@ class CreateTableFormPenilaianItem extends Migration
             $table->bigIncrements('id');
             $table->unsignedSmallInteger('form_penilaian_id');
             $table->string('nama', 60);
-            $table->integer('min')->nullable();
-            $table->integer('max')->nullable();
+            $table->smallInteger('min')->nullable();
+            $table->smallInteger('max')->nullable();
             $table->timestamps();
 
             $table->foreign('form_penilaian_id')
